+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This version of ONE adds support for loading .npz files.
 
 - one.alf.io.load_file_content loads npz files and returns only array if single compressed array with default name of 'arr_0'.
 - log warning when instantiating RegistrationClient with AlyxClient REST cache active
+- bugfix in load_collection when one or more files missing 
 
 ## [2.7.0]
 This version of ONE adds support for Alyx 2.0.0 and pandas 3.0.0 with dataset QC filters. This version no longer supports 'data' search filter.